chore: 로그인 유지 박스 기능 추가, 회원가입 클릭 시 회원가입 뷰로 전환
로그인 유지 박스를 활용하여, 새로 고침 시 로그인이 유지될지 안 될지를 설정할 수 있도록 변경하였고,
로그인창에서 회원가입 클릭 시 회원가입 뷰로 전환하도록 클릭 이벤트를 추가하였습니다.

+++ FE/src/components/LoginModal/LoginModal.js
@@ -14,6 +14,7 @@ import naver from "../../assets/naver.png";
 
 function LoginModal() {
   const [userId, setUserId] = useState("");
+  const [keepLogin, setKeepLogin] = useState(false);
   const [password, setPassword] = useState("");
   const setUserState = useSetRecoilState(userAtom);
   const setHeaderState = useSetRecoilState(headerAtom);
@@ -40,7 +41,9 @@ function LoginModal() {
             isLogin: true,
             accessToken: data.accessToken,
           }));
-          localStorage.setItem("accessToken", data.accessToken);
+          if (keepLogin) {
+            localStorage.setItem("accessToken", data.accessToken);
+          }
         } else {
           errorRef.current.innerText = data.message;
         }
@@ -87,7 +90,11 @@ function LoginModal() {
             onChange={(e) => setPassword(e.target.value)}
           />
           <CheckBar>
-            <input type='checkbox' />
+            <input
+              type='checkbox'
+              checked={keepLogin}
+              onChange={setKeepLogin}
+            />
             <div>로그인 유지</div>
           </CheckBar>
         </InputBar>
@@ -97,7 +104,15 @@ function LoginModal() {
             로그인
           </ModalButton>
         </ModalButtonContainer>
-        <HelpBar>
+        <HelpBar
+          onClick={() => {
+            setHeaderState((prev) => ({
+              ...prev,
+              isLogin: false,
+              isSignUp: true,
+            }));
+          }}
+        >
           <div>회원가입</div>
           <HelpBarBorder />
           <div>아이디/비밀번호 찾기</div>
@@ -165,6 +180,8 @@ const HelpBar = styled.div`
 
   font-size: 1rem;
   color: #ffffff;
+
+  cursor: pointer;
 `;
 
 const HelpBarBorder = styled.div`
